refactor: extract inners to canopus submodules
