+++ crates/canopus/src/canopus/validation.rs
@@ -4,7 +4,7 @@
 use crate::core::errors::{
     CodeownersValidationError, ConsistencyIssue, DiagnosticKind, StructuralIssue, ValidationDiagnostic,
 };
-use crate::core::models::codeowners::{CodeOwners, CodeOwnersAttributes, CodeOwnersEntry};
+use crate::core::models::codeowners::{CodeOwners, CodeOwnersContext, CodeOwnersEntry};
 use crate::core::models::handles::Owner;
 use crate::infra::github::{CheckGithubConsistency, GithubConsistencyChecker};
 use crate::infra::paths::{DirWalking, PathWalker};
@@ -21,7 +21,7 @@ pub struct CodeOwnersValidator {
 impl CodeOwnersValidator {
     pub async fn validate_codeowners(
         &self,
-        codeowners_file: CodeOwnersAttributes,
+        codeowners_file: CodeOwnersContext,
         organization_name: &str,
     ) -> anyhow::Result<()> {
         let project_root = codeowners_file.project_root.as_path();
@@ -257,5 +257,306 @@ impl CodeOwnersValidator {
             github_consistency_checker,
             path_walker,
         }
+    }
+}
+
+#[cfg(test)]
+mod test_builders {
+    use crate::canopus::validation::CodeOwnersValidator;
+    use crate::core::models::codeowners::CodeOwnersContext;
+    use crate::infra::github::{FakeGithubState, GithubConsistencyChecker};
+    use crate::infra::paths::PathWalker;
+    use std::path::PathBuf;
+
+    pub fn codeowners_attributes(contents: &str) -> CodeOwnersContext {
+        CodeOwnersContext {
+            project_root: PathBuf::from("/usr/projects/my-project"),
+            location: PathBuf::from("/usr/projects/my-project/.github/CODEOWNERS"),
+            contents: contents.to_string(),
+        }
+    }
+
+    pub fn structural_only_codeowners_validator(project_paths: Vec<&str>) -> CodeOwnersValidator {
+        let path_walker = PathWalker::with_paths(project_paths);
+        let consistency_checker = GithubConsistencyChecker::ConsistentState;
+        CodeOwnersValidator::new(consistency_checker, path_walker)
+    }
+
+    pub fn consistency_aware_codeowners_validator(
+        project_paths: Vec<&str>,
+        state: FakeGithubState,
+    ) -> CodeOwnersValidator {
+        let path_walker = PathWalker::with_paths(project_paths);
+        let consistency_checker = GithubConsistencyChecker::FakeChecks(state);
+        CodeOwnersValidator::new(consistency_checker, path_walker)
+    }
+}
+
+#[cfg(test)]
+mod structural_validation_tests {
+    use crate::canopus::validation::test_builders;
+    use crate::core::errors::test_helpers::DiagnosticKindFactory;
+    use crate::core::errors::{CodeownersValidationError, ValidationDiagnostic};
+    use assertor::{EqualityAssertion, ResultAssertion};
+    use indoc::indoc;
+
+    #[tokio::test]
+    async fn should_find_no_syntax_issues() {
+        let contents = indoc! {"
+            *.rs    @org/rustaceans
+        "};
+
+        let project_paths = vec!["main.rs"];
+
+        let context = test_builders::codeowners_attributes(contents);
+        let validator = test_builders::structural_only_codeowners_validator(project_paths);
+
+        let validation = validator.validate_codeowners(context, "dotanuki-labs").await;
+
+        assertor::assert_that!(validation).is_ok();
+    }
+
+    #[tokio::test]
+    async fn should_detect_owners_syntax_issue() {
+        let contents = indoc! {"
+            *.rs    org/rustaceans
+        "};
+
+        let context = test_builders::codeowners_attributes(contents);
+        let validator = test_builders::structural_only_codeowners_validator(vec![]);
+
+        let validation = validator.validate_codeowners(context, "dotanuki-labs").await;
+
+        let issue = ValidationDiagnostic::builder()
+            .kind(DiagnosticKindFactory::invalid_syntax())
+            .line_number(0)
+            .description("cannot parse owner")
+            .build();
+
+        let expected = CodeownersValidationError::from(issue);
+
+        assertor::assert_that!(validation.into()).is_equal_to(expected);
+    }
+
+    #[tokio::test]
+    async fn should_detect_glob_syntax_issue() {
+        let contents = indoc! {"
+            [z-a]*.rs    @org/crabbers
+        "};
+
+        let context = test_builders::codeowners_attributes(contents);
+        let validator = test_builders::structural_only_codeowners_validator(vec![]);
+
+        let validation = validator.validate_codeowners(context, "dotanuki-labs").await;
+
+        let issue = ValidationDiagnostic::builder()
+            .kind(DiagnosticKindFactory::invalid_syntax())
+            .line_number(0)
+            .description("invalid glob pattern")
+            .build();
+
+        let expected = CodeownersValidationError::from(issue);
+
+        assertor::assert_that!(validation.into()).is_equal_to(expected);
+    }
+
+    #[tokio::test]
+    async fn should_report_multiple_issues_for_the_same_entry() {
+        let contents = indoc! {"
+            [z-a]*.rs    org/crabbers
+        "};
+
+        let context = test_builders::codeowners_attributes(contents);
+        let validator = test_builders::structural_only_codeowners_validator(vec![]);
+
+        let validation = validator.validate_codeowners(context, "dotanuki-labs").await;
+
+        let invalid_glob = ValidationDiagnostic::builder()
+            .kind(DiagnosticKindFactory::invalid_syntax())
+            .line_number(0)
+            .description("invalid glob pattern")
+            .build();
+
+        let invalid_owner = ValidationDiagnostic::builder()
+            .kind(DiagnosticKindFactory::invalid_syntax())
+            .line_number(0)
+            .description("cannot parse owner")
+            .build();
+
+        let expected = CodeownersValidationError::with(vec![invalid_glob, invalid_owner]);
+
+        assertor::assert_that!(validation.into()).is_equal_to(expected);
+    }
+
+    #[tokio::test]
+    async fn should_detect_dangling_globs() {
+        let contents = indoc! {"
+            *.rs            @org/rustaceans
+            .automation/    @org/infra
+        "};
+
+        let project_paths = vec!["validation.rs"];
+
+        let context = test_builders::codeowners_attributes(contents);
+        let validator = test_builders::structural_only_codeowners_validator(project_paths);
+
+        let validation = validator.validate_codeowners(context, "dotanuki-labs").await;
+
+        let issue = ValidationDiagnostic::builder()
+            .kind(DiagnosticKindFactory::dangling_glob_pattern())
+            .line_number(1)
+            .description(".automation/ does not match any project path")
+            .build();
+
+        let expected = CodeownersValidationError::from(issue);
+
+        assertor::assert_that!(validation.into()).is_equal_to(expected);
+    }
+
+    #[tokio::test]
+    async fn should_detect_strictly_duplicated_ownership_rules() {
+        let contents = indoc! {"
+            *.rs     @org/rustaceans
+            docs/    @org/rustaceans
+            *.rs     @org/crabbers @ubiratansoares
+        "};
+
+        let project_paths = vec!["validation.rs", "docs/", "docs/README.md"];
+
+        let context = test_builders::codeowners_attributes(contents);
+        let validator = test_builders::structural_only_codeowners_validator(project_paths);
+
+        let validation = validator.validate_codeowners(context, "dotanuki-labs").await;
+
+        let duplicated_ownership = ValidationDiagnostic::builder()
+            .kind(DiagnosticKindFactory::duplicate_ownership())
+            .line_number(0)
+            .description("*.rs defined multiple times : lines [0, 2]")
+            .build();
+
+        let expected = CodeownersValidationError::from(duplicated_ownership);
+
+        assertor::assert_that!(validation.into()).is_equal_to(expected);
+    }
+
+    #[tokio::test]
+    async fn should_detect_multiple_non_syntax_issues() {
+        let contents = indoc! {"
+            *.rs        @org/rustaceans
+            docs/       @org/devs
+            *.rs        @org/crabbers @ubiratansoares
+        "};
+
+        let project_paths = vec!["validation.rs", ".github/", ".github/CODEOWNERS"];
+
+        let context = test_builders::codeowners_attributes(contents);
+        let validator = test_builders::structural_only_codeowners_validator(project_paths);
+
+        let validation = validator.validate_codeowners(context, "dotanuki-labs").await;
+
+        let dangling_glob = ValidationDiagnostic::builder()
+            .kind(DiagnosticKindFactory::dangling_glob_pattern())
+            .line_number(1)
+            .description("docs/ does not match any project path")
+            .build();
+
+        let duplicated_ownership = ValidationDiagnostic::builder()
+            .kind(DiagnosticKindFactory::duplicate_ownership())
+            .line_number(0)
+            .description("*.rs defined multiple times : lines [0, 2]")
+            .build();
+
+        let expected = CodeownersValidationError::with(vec![dangling_glob, duplicated_ownership]);
+        assertor::assert_that!(validation.into()).is_equal_to(expected);
+    }
+}
+
+#[cfg(test)]
+mod consistency_validation_tests {
+    use crate::canopus::validation::test_builders;
+    use crate::core::errors::test_helpers::DiagnosticKindFactory;
+    use crate::core::errors::{CodeownersValidationError, ValidationDiagnostic};
+    use crate::infra::github;
+    use assertor::{EqualityAssertion, ResultAssertion};
+    use indoc::indoc;
+
+    #[tokio::test]
+    async fn should_find_no_consistency_issues() {
+        let contents = indoc! {"
+            *.rs        @dotanuki-labs/rustaceans
+            .github/    @ubiratansoares
+        "};
+
+        let project_paths = vec![".github/", "main.rs"];
+
+        let github_state = github::FakeGithubState::builder()
+            .add_known_user("@ubiratansoares")
+            .add_known_team("@dotanuki-labs/rustaceans")
+            .build();
+
+        let context = test_builders::codeowners_attributes(contents);
+        let validator = test_builders::consistency_aware_codeowners_validator(project_paths, github_state);
+
+        let validation = validator.validate_codeowners(context, "dotanuki-labs").await;
+
+        assertor::assert_that!(validation).is_ok();
+    }
+
+    #[tokio::test]
+    async fn should_detect_non_existing_github_user() {
+        let contents = indoc! {"
+            *.rs        @dotanuki-labs/rustaceans
+            .github/    @ufs
+        "};
+
+        let project_paths = vec![".github/", "main.rs"];
+
+        let github_state = github::FakeGithubState::builder()
+            .add_known_team("@dotanuki-labs/rustaceans")
+            .build();
+
+        let context = test_builders::codeowners_attributes(contents);
+        let validator = test_builders::consistency_aware_codeowners_validator(project_paths, github_state);
+
+        let validation = validator.validate_codeowners(context, "dotanuki-labs").await;
+
+        let user_not_found = ValidationDiagnostic::builder()
+            .kind(DiagnosticKindFactory::user_does_not_belong_to_organization("ufs"))
+            .line_number(1)
+            .description("'ufs' user does not belong to this organization")
+            .build();
+
+        let expected = CodeownersValidationError::from(user_not_found);
+        assertor::assert_that!(validation.into()).is_equal_to(expected);
+    }
+
+    #[tokio::test]
+    async fn should_detect_non_existing_github_team() {
+        let contents = indoc! {"
+            *.rs        @dotanuki-labs/rustaceans
+            docs/       @dotanuki-labs/writers
+            .github/    @dotanuki-labs/devops
+        "};
+
+        let project_paths = vec![".github/", "docs/", "main.rs"];
+
+        let github_state = github::FakeGithubState::builder()
+            .add_known_team("@dotanuki-labs/rustaceans")
+            .add_known_team("@dotanuki-labs/writers")
+            .build();
+
+        let context = test_builders::codeowners_attributes(contents);
+        let validator = test_builders::consistency_aware_codeowners_validator(project_paths, github_state);
+
+        let validation = validator.validate_codeowners(context, "dotanuki-labs").await;
+
+        let user_not_found = ValidationDiagnostic::builder()
+            .kind(DiagnosticKindFactory::team_does_not_exist("dotanuki-labs", "devops"))
+            .line_number(2)
+            .description("'devops' team does not belong to 'dotanuki-labs' organization")
+            .build();
+
+        let expected = CodeownersValidationError::from(user_not_found);
+        assertor::assert_that!(validation.into()).is_equal_to(expected);
     }
 }
